+++ lib/onedrive_api.dart
@@ -1,4 +1,6 @@
 library;
 
 export 'src/params.dart';
-export 'src/create_folder.dart' show createFolder, Response;
+export 'src/create_folder.dart' show createFolder;
+export 'src/delete.dart' show delete;
+export 'src/core.dart' show Response;