+++ test/delete_test.dart
@@ -5,27 +5,22 @@ import 'package:test/test.dart';
 void main() async {
   var env = DotEnv()..load();
   String accessToken = env['ACCESS_TOKEN']!;
-  group('Create Folder Tests', () {
-    test('Random folder at root', () async {
+  group('Delete Item Tests', () {
+    test('Delete Item', () async {
       FolderParams folderParams = FolderParams();
       folderParams.accessToken = accessToken;
-      folderParams.name = "Test Folder";
+      folderParams.name = "Test Folder 2";
 
       Response response = await createFolder(folderParams);
       expect(response.statusCode, 201);
-      expect(response.itemId, isNotNull);
-      expect(response.errorMsg, isNull);
 
-      String createdItem = response.itemId!;
-
-      response = await createFolder(folderParams);
-      expect(response.statusCode, 409);
-      expect(response.itemId, isNull);
-      expect(response.errorMsg, isNotNull);
-
-      folderParams.itemId = createdItem;
+      folderParams.itemId = response.itemId;
       response = await delete(folderParams);
       expect(response.statusCode, 204);
+
+      response = await delete(folderParams);
+      expect(response.statusCode, 404);
+      expect(response.errorMsg, isNotNull);
     });
   });
 }