+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.1.0
+- Adds folder creation endpoint.
 
-- Initial release.
-- Adds support for folder creation endpoint.
+## 0.2.0
+- Adds item deletion endpoint.